+++ tests/Mock/Command/Output.php
@@ -9,7 +9,10 @@ class Output implements OutputInterface
 {
     private $messages = [];
 
+    /** @var OutputFormatterInterface */
     private $formatter;
+
+    private $verbosity = self::VERBOSITY_NORMAL;
 
     public function getFormatter()
     {
@@ -18,28 +21,27 @@ class Output implements OutputInterface
 
     public function getVerbosity()
     {
-
+        return $this->verbosity;
     }
 
     public function isDecorated()
     {
-
+        return $this->formatter->isDecorated();
     }
 
     public function setDecorated($decorated)
     {
-
+        $this->formatter->setDecorated($decorated);
     }
 
     public function setFormatter(OutputFormatterInterface $formatter)
     {
         $this->formatter = $formatter;
-        return $this;
     }
 
     public function setVerbosity($level)
     {
-
+        $this->verbosity = $level;
     }
 
     public function write($messages, $newline = false, $options = 0)
@@ -73,21 +75,21 @@ class Output implements OutputInterface
 
     public function isDebug()
     {
-
+        return self::VERBOSITY_DEBUG <= $this->verbosity;
     }
 
     public function isQuiet()
     {
-
+        return self::VERBOSITY_QUIET === $this->verbosity;
     }
 
     public function isVerbose()
     {
-
+        return self::VERBOSITY_VERBOSE <= $this->verbosity;
     }
 
     public function isVeryVerbose()
     {
-
+        return self::VERBOSITY_VERY_VERBOSE <= $this->verbosity;
     }
 }